Use size_t instead of int to store bit index for peers
Use size_t instead of int to store bit index for peers in RibOut,
RibOutUpdates, UpdateQueue and BgpUpdateSender.

Partial-Bug: 1602347

+++ src/bgp/bgp_ribout_updates.cc
@@ -666,12 +666,12 @@ size_t RibOutUpdates::queue_marker_count(int queue_id) const {
     return queue->marker_count();
 }
 
-bool RibOutUpdates::QueueJoin(int queue_id, int bit) {
+bool RibOutUpdates::QueueJoin(int queue_id, size_t bit) {
     UpdateQueue *queue = queue_vec_[queue_id];
     return queue->Join(bit);
 }
 
-void RibOutUpdates::QueueLeave(int queue_id, int bit) {
+void RibOutUpdates::QueueLeave(int queue_id, size_t bit) {
     UpdateQueue *queue = queue_vec_[queue_id];
     queue->Leave(bit);
 }
